+++ rockset/api/collections_api.py
@@ -39,6 +39,7 @@ from rockset.model.kinesis_collection_creation_request import KinesisCollectionC
 from rockset.model.list_collections_response import ListCollectionsResponse
 from rockset.model.mongodb_collection_creation_request import MongodbCollectionCreationRequest
 from rockset.model.s3_collection_creation_request import S3CollectionCreationRequest
+from rockset.model.snowflake_collection_creation_request import SnowflakeCollectionCreationRequest
 from rockset.models import *
 
 
@@ -609,6 +610,63 @@ class Collections(object):
                 'location_map': {
                     'workspace': 'path',
                     's3_collection_creation_request': 'body',
+                },
+                'collection_format_map': {
+                }
+            },
+            headers_map={
+                'accept': [
+                    'application/json'
+                ],
+                'content_type': [
+                    'application/json'
+                ]
+            },
+            api_client=api_client
+        )
+        self.create_snowflake_collection_endpoint = _Endpoint(
+            settings={
+                'response_type': (CreateCollectionResponse,),
+                'auth': [
+                    'apikey'
+                ],
+                'endpoint_path': '/v1/orgs/self/ws/{workspace}/collections',
+                'operation_id': 'create_snowflake_collection',
+                'http_method': 'POST',
+                'servers': None,
+            },
+            params_map={
+                'all': [
+                    'workspace',
+                    'snowflake_collection_creation_request',
+                ],
+                'required': [
+                    'workspace',
+                ],
+                'nullable': [
+                ],
+                'enum': [
+                ],
+                'validation': [
+                ]
+            },
+            root_map={
+                'validations': {
+                },
+                'allowed_values': {
+                },
+                'openapi_types': {
+                    'workspace':
+                        (str,),
+                    'snowflake_collection_creation_request':
+                        (SnowflakeCollectionCreationRequest,),
+                },
+                'attribute_map': {
+                    'workspace': 'workspace',
+                },
+                'location_map': {
+                    'workspace': 'path',
+                    'snowflake_collection_creation_request': 'body',
                 },
                 'collection_format_map': {
                 }
@@ -2576,6 +2634,181 @@ class Collections(object):
             kwargs['s3_collection_creation_request']
         return self.create_s3_collection_endpoint.call_with_http_info(**kwargs)
 
+    def create_snowflake_collection(
+        self,
+        *,
+        name: str,
+        clustering_key: typing.Sequence[FieldPartition] = None,
+        description: str = None,
+        event_time_info: EventTimeInfo = None,
+        field_mapping_query: FieldMappingQuery = None,
+        field_mappings: typing.Sequence[FieldMappingV2] = None,
+        insert_only: bool = None,
+        retention_secs: int = None,
+        sources: typing.Sequence[SnowflakeSourceWrapper] = None,
+        workspace = "commons",
+        **kwargs
+    ) -> typing.Union[CreateCollectionResponse, asyncio.Future]:
+        """Create snowflake collection  # noqa: E501
+
+        Create new collection in a workspace.  # noqa: E501
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async_req=True
+
+        ```python
+        rs = RocksetClient(api_key=APIKEY)
+        future = rs.Collections.create_snowflake_collection(
+            clustering_key=[
+                FieldPartition(
+                    field_name="address.city.zipcode",
+                    keys=["value1","value2"],
+                    type="AUTO",
+                ),
+            ],
+            description="transactions from stores worldwide",
+            event_time_info=EventTimeInfo(
+                field="timestamp",
+                format="seconds_since_epoch",
+                time_zone="UTC",
+            ),
+            field_mapping_query=FieldMappingQuery(
+                sql="sql",
+            ),
+            field_mappings=[
+                FieldMappingV2(
+                    input_fields=[
+                        InputField(
+                            field_name="address.city.zipcode",
+                            if_missing="SKIP",
+                            is_drop=True,
+                            param="zip",
+                        ),
+                    ],
+                    name="myTestMapping",
+                    output_field=OutputField(
+                        field_name="zip_hash",
+                        on_error="SKIP",
+                        value=SqlExpression(
+                            sql="SHA256()",
+                        ),
+                    ),
+                ),
+            ],
+            insert_only=True,
+            name="global-transactions",
+            retention_secs=1000000,
+            sources=[
+                SnowflakeSourceWrapper(
+                    format_params=FormatParams(
+                        csv=CsvParams(
+                            column_names=["c1","c2","c3"],
+                            column_types=["BOOLEAN","INTEGER","FLOAT","STRING"],
+                            encoding="UTF-8",
+                            escape_char="\\",
+                            first_line_as_column_names=True,
+                            quote_char="\"",
+                            separator=",",
+                        ),
+                        json=True,
+                        mssql_dms=True,
+                        mysql_dms=True,
+                        oracle_dms=True,
+                        postgres_dms=True,
+                        xml=XmlParams(
+                            attribute_prefix="_attr",
+                            doc_tag="row",
+                            encoding="UTF-8",
+                            root_tag="root",
+                            value_tag="value",
+                        ),
+                    ),
+                    integration_name="aws-integration",
+                    database="NASDAQ",
+                    schema="PUBLIC",
+                    table_name="COMPANIES",
+                    warehouse="COMPUTE_XL",
+                ),
+            ],
+            async_req=True,
+        )
+        result = await future
+        ```
+
+        Keyword Args:
+            workspace (str): name of the workspace. [required] if omitted the server will use the default value of "commons"
+            clustering_key ([FieldPartition]): list of clustering fields. [optional]
+            description (str): text describing the collection. [optional]
+            event_time_info (EventTimeInfo): [optional]
+            field_mapping_query (FieldMappingQuery): [optional]
+            field_mappings ([FieldMappingV2]): list of mappings. [optional]
+            insert_only (bool): If true disallows updates and deletes, but makes indexing more efficient. [optional]
+            name (str): unique identifier for collection, can contain alphanumeric or dash characters. [required]
+            retention_secs (int): number of seconds after which data is purged, based on event time. [optional]
+            sources ([SnowflakeSourceWrapper]): List of sources from which to ingest data. [optional]
+            _return_http_data_only (bool): response data without head status
+                code and headers. Default is True.
+            _preload_content (bool): if False, the urllib3.HTTPResponse object
+                will be returned without reading/decoding response data.
+                Default is True.
+            _request_timeout (int/float/tuple): timeout setting for this request. If
+                one number provided, it will be total request timeout. It can also
+                be a pair (tuple) of (connection, read) timeouts.
+                Default is None.
+            _check_input_type (bool): specifies if type checking
+                should be done one the data sent to the server.
+                Default is True.
+            _check_return_type (bool): specifies if type checking
+                should be done on the data received from the server.
+                If False, the client will also not convert nested inner objects
+                into the respective model types (the outermost object
+                is still converted to the model).
+                Default is True.
+            _spec_property_naming (bool): True if the variable names in the input data
+                are serialized names, as specified in the OpenAPI document.
+                False if the variable names in the input data
+                are pythonic names, e.g. snake case (default)
+            _content_type (str/None): force body content-type.
+                Default is None and content-type will be predicted by allowed
+                content-types and body.
+            _host_index (int/None): specifies the index of the server
+                that we want to use.
+                Default is read from the configuration.
+            async_req (bool): execute request asynchronously
+
+        Returns:
+            CreateCollectionResponse
+                If the method is called asynchronously, returns an asyncio.Future which resolves to the response.
+        """
+        kwargs['async_req'] = kwargs.get(
+            'async_req', False
+        )
+        kwargs['_return_http_data_only'] = kwargs.get(
+            '_return_http_data_only', True
+        )
+        kwargs['_preload_content'] = kwargs.get(
+            '_preload_content', True
+        )
+        kwargs['_request_timeout'] = kwargs.get(
+            '_request_timeout', None
+        )
+        kwargs['_check_input_type'] = kwargs.get(
+            '_check_input_type', True
+        )
+        kwargs['_check_return_type'] = kwargs.get(
+            '_check_return_type', True
+        )
+        kwargs['_spec_property_naming'] = kwargs.get(
+            '_spec_property_naming', False
+        )
+        kwargs['_content_type'] = kwargs.get(
+            '_content_type')
+        kwargs['_host_index'] = kwargs.get('_host_index')
+        kwargs['workspace'] = \
+            workspace
+        kwargs['snowflake_collection_creation_request'] = \
+            kwargs['snowflake_collection_creation_request']
+        return self.create_snowflake_collection_endpoint.call_with_http_info(**kwargs)
+
     def delete(
         self,
         *,
@@ -2940,3 +3173,5 @@ class Collections(object):
     return_types_dict['create_mongodb_collection'] = MongodbCollectionCreationRequest
     body_params_dict['create_s3_collection'] = 's3_collection_creation_request'
     return_types_dict['create_s3_collection'] = S3CollectionCreationRequest
+    body_params_dict['create_snowflake_collection'] = 'snowflake_collection_creation_request'
+    return_types_dict['create_snowflake_collection'] = SnowflakeCollectionCreationRequest