+++ rockset/api/integrations_api.py
@@ -38,6 +38,7 @@ from rockset.model.list_integrations_response import ListIntegrationsResponse
 from rockset.model.mongodb_integration_creation_request import MongodbIntegrationCreationRequest
 from rockset.model.s3_integration_creation_request import S3IntegrationCreationRequest
 from rockset.model.segment_integration_creation_request import SegmentIntegrationCreationRequest
+from rockset.model.snowflake_integration_creation_request import SnowflakeIntegrationCreationRequest
 from rockset.models import *
 
 
@@ -497,6 +498,57 @@ class Integrations(object):
                 },
                 'location_map': {
                     'segment_integration_creation_request': 'body',
+                },
+                'collection_format_map': {
+                }
+            },
+            headers_map={
+                'accept': [
+                    'application/json'
+                ],
+                'content_type': [
+                    'application/json'
+                ]
+            },
+            api_client=api_client
+        )
+        self.create_snowflake_integration_endpoint = _Endpoint(
+            settings={
+                'response_type': (CreateIntegrationResponse,),
+                'auth': [
+                    'apikey'
+                ],
+                'endpoint_path': '/v1/orgs/self/integrations',
+                'operation_id': 'create_snowflake_integration',
+                'http_method': 'POST',
+                'servers': None,
+            },
+            params_map={
+                'all': [
+                    'snowflake_integration_creation_request',
+                ],
+                'required': [
+                ],
+                'nullable': [
+                ],
+                'enum': [
+                ],
+                'validation': [
+                ]
+            },
+            root_map={
+                'validations': {
+                },
+                'allowed_values': {
+                },
+                'openapi_types': {
+                    'snowflake_integration_creation_request':
+                        (SnowflakeIntegrationCreationRequest,),
+                },
+                'attribute_map': {
+                },
+                'location_map': {
+                    'snowflake_integration_creation_request': 'body',
                 },
                 'collection_format_map': {
                 }
@@ -1528,6 +1580,111 @@ class Integrations(object):
             kwargs['segment_integration_creation_request']
         return self.create_segment_integration_endpoint.call_with_http_info(**kwargs)
 
+    def create_snowflake_integration(
+        self,
+        *,
+        name: str,
+        description: str = None,
+        snowflake: SnowflakeIntegration = None,
+        **kwargs
+    ) -> typing.Union[CreateIntegrationResponse, asyncio.Future]:
+        """Create snowflake integration  # noqa: E501
+
+        Create a new integration.  # noqa: E501
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async_req=True
+
+        ```python
+        rs = RocksetClient(api_key=APIKEY)
+        future = rs.Integrations.create_snowflake_integration(
+            description="AWS account with event data for the data science team.",
+            name="event-logs",
+            snowflake=SnowflakeIntegration(
+                aws_access_key=AwsAccessKey(
+                    aws_access_key_id="AKIAIOSFODNN7EXAMPLE",
+                    aws_secret_access_key="wJal....",
+                ),
+                aws_role=AwsRole(
+                    aws_role_arn="arn:aws:iam::2378964092:role/rockset-role",
+                ),
+                default_warehouse="default_warehouse_example",
+                password="password_example",
+                s3_export_path="s3://bucket/prefix",
+                snowflake_url="acme-marketing-test-account.snowflakecomputing.com",
+                user_role="user_role_example",
+                username="username_example",
+            ),
+            async_req=True,
+        )
+        result = await future
+        ```
+
+        Keyword Args:
+            description (str): longer explanation for the integration. [optional]
+            name (str): descriptive label. [required]
+            snowflake (SnowflakeIntegration): [optional]
+            _return_http_data_only (bool): response data without head status
+                code and headers. Default is True.
+            _preload_content (bool): if False, the urllib3.HTTPResponse object
+                will be returned without reading/decoding response data.
+                Default is True.
+            _request_timeout (int/float/tuple): timeout setting for this request. If
+                one number provided, it will be total request timeout. It can also
+                be a pair (tuple) of (connection, read) timeouts.
+                Default is None.
+            _check_input_type (bool): specifies if type checking
+                should be done one the data sent to the server.
+                Default is True.
+            _check_return_type (bool): specifies if type checking
+                should be done on the data received from the server.
+                If False, the client will also not convert nested inner objects
+                into the respective model types (the outermost object
+                is still converted to the model).
+                Default is True.
+            _spec_property_naming (bool): True if the variable names in the input data
+                are serialized names, as specified in the OpenAPI document.
+                False if the variable names in the input data
+                are pythonic names, e.g. snake case (default)
+            _content_type (str/None): force body content-type.
+                Default is None and content-type will be predicted by allowed
+                content-types and body.
+            _host_index (int/None): specifies the index of the server
+                that we want to use.
+                Default is read from the configuration.
+            async_req (bool): execute request asynchronously
+
+        Returns:
+            CreateIntegrationResponse
+                If the method is called asynchronously, returns an asyncio.Future which resolves to the response.
+        """
+        kwargs['async_req'] = kwargs.get(
+            'async_req', False
+        )
+        kwargs['_return_http_data_only'] = kwargs.get(
+            '_return_http_data_only', True
+        )
+        kwargs['_preload_content'] = kwargs.get(
+            '_preload_content', True
+        )
+        kwargs['_request_timeout'] = kwargs.get(
+            '_request_timeout', None
+        )
+        kwargs['_check_input_type'] = kwargs.get(
+            '_check_input_type', True
+        )
+        kwargs['_check_return_type'] = kwargs.get(
+            '_check_return_type', True
+        )
+        kwargs['_spec_property_naming'] = kwargs.get(
+            '_spec_property_naming', False
+        )
+        kwargs['_content_type'] = kwargs.get(
+            '_content_type')
+        kwargs['_host_index'] = kwargs.get('_host_index')
+        kwargs['snowflake_integration_creation_request'] = \
+            kwargs['snowflake_integration_creation_request']
+        return self.create_snowflake_integration_endpoint.call_with_http_info(**kwargs)
+
     def delete(
         self,
         *,
@@ -1798,3 +1955,5 @@ class Integrations(object):
     return_types_dict['create_s3_integration'] = S3IntegrationCreationRequest
     body_params_dict['create_segment_integration'] = 'segment_integration_creation_request'
     return_types_dict['create_segment_integration'] = SegmentIntegrationCreationRequest
+    body_params_dict['create_snowflake_integration'] = 'snowflake_integration_creation_request'
+    return_types_dict['create_snowflake_integration'] = SnowflakeIntegrationCreationRequest