+++ rockset/model/create_integration_request.py
@@ -39,6 +39,7 @@ def lazy_import():
     from rockset.model.mongo_db_integration import MongoDbIntegration
     from rockset.model.s3_integration import S3Integration
     from rockset.model.segment_integration import SegmentIntegration
+    from rockset.model.snowflake_integration import SnowflakeIntegration
     globals()['AzureBlobStorageIntegration'] = AzureBlobStorageIntegration
     globals()['AzureEventHubsIntegration'] = AzureEventHubsIntegration
     globals()['DynamodbIntegration'] = DynamodbIntegration
@@ -48,6 +49,7 @@ def lazy_import():
     globals()['MongoDbIntegration'] = MongoDbIntegration
     globals()['S3Integration'] = S3Integration
     globals()['SegmentIntegration'] = SegmentIntegration
+    globals()['SnowflakeIntegration'] = SnowflakeIntegration
 
 
 class CreateIntegrationRequest(ModelNormal):
@@ -114,6 +116,7 @@ class CreateIntegrationRequest(ModelNormal):
             'mongodb': (MongoDbIntegration, none_type),  # noqa: E501
             's3': (S3Integration, none_type),  # noqa: E501
             'segment': (SegmentIntegration, none_type),  # noqa: E501
+            'snowflake': (SnowflakeIntegration, none_type),  # noqa: E501
         }
 
     @cached_property
@@ -133,6 +136,7 @@ class CreateIntegrationRequest(ModelNormal):
         'mongodb': 'mongodb',  # noqa: E501
         's3': 's3',  # noqa: E501
         'segment': 'segment',  # noqa: E501
+        'snowflake': 'snowflake',  # noqa: E501
     }
 
     read_only_vars = {
@@ -189,6 +193,7 @@ class CreateIntegrationRequest(ModelNormal):
             mongodb (MongoDbIntegration): [optional]  # noqa: E501
             s3 (S3Integration): [optional]  # noqa: E501
             segment (SegmentIntegration): [optional]  # noqa: E501
+            snowflake (SnowflakeIntegration): [optional]  # noqa: E501
         """
 
         _check_type = kwargs.pop('_check_type', True)
@@ -252,6 +257,7 @@ class CreateIntegrationRequest(ModelNormal):
             mongodb (MongoDbIntegration): [optional]  # noqa: E501
             s3 (S3Integration): [optional]  # noqa: E501
             segment (SegmentIntegration): [optional]  # noqa: E501
+            snowflake (SnowflakeIntegration): [optional]  # noqa: E501
             _check_type (bool): if True, values for parameters in openapi_types
                                 will be type checked and a TypeError will be
                                 raised if the wrong type is input.