+++ test/test_collections_api.py
@@ -823,6 +823,87 @@ def test_create_s3_collection(get_client, mock_request, request_validator):
             validate_call(e, request_validator)
 
 
+def test_create_snowflake_collection(get_client, mock_request, request_validator):
+    with mock_request:
+        rs = get_client
+        try:
+            rs.Collections.create_snowflake_collection(
+                clustering_key=[
+                    FieldPartition(
+                        field_name="address.city.zipcode",
+                        keys=["value1", "value2"],
+                        type="AUTO",
+                    ),
+                ],
+                description="transactions from stores worldwide",
+                event_time_info=EventTimeInfo(
+                    field="timestamp",
+                    format="seconds_since_epoch",
+                    time_zone="UTC",
+                ),
+                field_mapping_query=FieldMappingQuery(
+                    sql="sql",
+                ),
+                field_mappings=[
+                    FieldMappingV2(
+                        input_fields=[
+                            InputField(
+                                field_name="address.city.zipcode",
+                                if_missing="SKIP",
+                                is_drop=True,
+                                param="zip",
+                            ),
+                        ],
+                        name="myTestMapping",
+                        output_field=OutputField(
+                            field_name="zip_hash",
+                            on_error="SKIP",
+                            value=SqlExpression(
+                                sql="SHA256()",
+                            ),
+                        ),
+                    ),
+                ],
+                insert_only=True,
+                name="global-transactions",
+                retention_secs=1000000,
+                sources=[
+                    SnowflakeSourceWrapper(
+                        format_params=FormatParams(
+                            csv=CsvParams(
+                                column_names=["c1", "c2", "c3"],
+                                column_types=["BOOLEAN", "INTEGER", "FLOAT", "STRING"],
+                                encoding="UTF-8",
+                                escape_char="\\",
+                                first_line_as_column_names=True,
+                                quote_char='"',
+                                separator=",",
+                            ),
+                            json=True,
+                            mssql_dms=True,
+                            mysql_dms=True,
+                            oracle_dms=True,
+                            postgres_dms=True,
+                            xml=XmlParams(
+                                attribute_prefix="_attr",
+                                doc_tag="row",
+                                encoding="UTF-8",
+                                root_tag="root",
+                                value_tag="value",
+                            ),
+                        ),
+                        integration_name="aws-integration",
+                        database="NASDAQ",
+                        schema="PUBLIC",
+                        table_name="COMPANIES",
+                        warehouse="COMPUTE_XL",
+                    ),
+                ],
+            )
+        except EarlyExit as e:
+            validate_call(e, request_validator)
+
+
 def test_delete(get_client, mock_request, request_validator):
     with mock_request:
         rs = get_client